Moved the debug classes to their own package as to not clutter the production code.
